PreparedStatement -> Statement
Kinda still batch execute, but we commit statement executes in batches. For statements, the changes will not be applied if no commits happen. Usually auto-commit is set to true so each execute will update the database, by setting it to false and manually committing in batches, we can also do the same as executeBatch()